lab9/practice5/requirements.txt
# this file defines the pip packages neede by this application
flask==3.0.2
azure-storage-blob==12.3.1
azure-cosmos
pika==1.2.0
python-dotenv==0.19.0
